C: Fix heli_bitmap_inner.horz_offset u8 → s8
Single usage site was already casting (s8) before arithmetic; the comment
said "might be s8". Change confirmed by same pattern as horz_clip.

+++ C/ChaseHQ/ChaseHQ.c
@@ -7386,7 +7386,7 @@ static void draw_helicoper_part(chqstate_t                *state,
   BC = state->dhl_helipos;
   state->doc_col_pos = -A; // in draw_object_common
 
-  screen_pos = (s8) DEinnerbitmap->horz_offset + BC; // loads byte and widens
+  screen_pos = DEinnerbitmap->horz_offset + BC;
 
 
   HLbitmap  = &DEinnerbitmap->bm; // Conv: Ops shuffled a bit
